backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
PyJWT[crypto]>=2.10.1
python-dotenv>=1.0.1
